backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers
dj-rest-auth
django-allauth
requests
django-extensions

